+++ ch5/queues/test-LinkedQueue.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+g++ test-LinkedQueue.cpp -o test-LinkedQueue
+./test-LinkedQueue
+
+rm test-LinkedQueue